+++ R/stat-poly-eq.R
@@ -37,8 +37,14 @@
 #'   for fits with \code{lm()} and \code{rlm()}, see `level` to control.)
 #' @param fm.values logical Add metadata and parameter estimates extracted from
 #'   the fitted model object; \code{FALSE} by default.
-#' @param fullrange Should the fit span the full range of the plot, or just the
-#'   range of the data group used in each fit?
+#' @param fullrange logical Should the fit prediction span the full
+#'   range of the plot, or just the range of the explanatory variable?
+#' @param limit.to character or numeric If character one of \code{""}, \code{"x"}, \code{"y"} or
+#'   \code{"xy"}. Should the fit prediction be constrained to the range of the
+#'   variables mapped to \code{x} and/or \code{y} in each data group? If numeric,
+#'   the \emph{new data} values to use for the explanatory variable when
+#'   computing the predicted line and confidence band. When
+#'   set \code{limit.to} silently overrides \code{fullrange}!
 #' @param level Level of confidence interval to use (0.95 by default).
 #' @param n Number of points at which to predict with the fitted model.
 #'
@@ -198,7 +204,24 @@
 #'   positioning using a vector of length > 1 for \code{label.x} and/or
 #'   \code{label.y} is the currently available workaround.
 #'
-##' @section Model fit methods supported:
+#' @section Range of the prediction line: The range of the prediction line is
+#'   controlled by parameters \code{fullrange} and \code{limit.to}.
+#'   \code{fullrange} is backwards compatible both with earlier versions of
+#'   'ggpmisc' and with \code{stat_smooth()} from 'ggplot2'; an argument passed
+#'   to \code{limit.to} overrides \code{fullrange} making it possible to
+#'   constrain the range to that of \code{x}, \code{y}, or both simultaneously,
+#'   with \code{"x"}, \code{"y"}, or \code{"xy"}, respectively, as argument.
+#'   \code{limit.to} also accepts a numeric vector of values to be used as
+#'   \code{newdata} when computing the prediction. Limiting the range based on
+#'   both aesthetics is the best approach for major axis regression (MA, SMA,
+#'   RMA) but can occasionally be useful also with some other methods when
+#'   slopes are very steep and error variance in the explanatory variable is
+#'   large. A numeric vector can be used to predict the response at specific
+#'   values of the explanatory variable. If a single or very few values are
+#'   predicted, it can be necessary to override the default \code{geom =
+#'   "smooth"} with \code{geom = "pointrange"}.
+#'
+#' @section Model fit methods supported:
 #'   Several model fit functions are supported explicitly (see tables), and some
 #'   of their differences smoothed out. Compatibility is checked late, based on
 #'   the class of the returned fitted model object. This makes it possible to
@@ -764,7 +787,8 @@ poly_eq_compute_group_fun <- function(data,
 
   if (exists("grp.label", data)) {
     if (length(unique(data[["grp.label"]])) > 1L) {
-      warning("Non-unique value in 'data$grp.label' using group index ", data[["group"]][1], " as label.")
+      warning("Non-unique value in 'data$grp.label' using group index ",
+              data[["group"]][1], " as label.")
       grp.label <- as.character(data[["group"]][1])
     } else {
       grp.label <- data[["grp.label"]][1]
@@ -823,7 +847,7 @@ poly_eq_compute_group_fun <- function(data,
   }
 
   # allow model formula selection by the model fit method
-  # extract formula from fitted model if possible, but fall back on argument if needed
+  # extract formula from fitted model object, fall back on argument on failure
   formula.ls <- fail_safe_formula(fm, method.args, verbose = TRUE)
 
   if ("fstatistic" %in% names(fm.summary)) {
@@ -914,7 +938,8 @@ poly_eq_compute_group_fun <- function(data,
   selector <- !is.na(coefs)
   coefs <- coefs[selector]
   if (!all(selector)) {
-    message("Terms dropped from model (singularity); n = ", nrow(data), " in group.")
+    message("Terms dropped from model (singularity); n = ",
+            nrow(data), " in group.")
   }
   # z is the object to be returned, i.e., passed to the geometry function
   # it must be a data.frame, and here we use one row per group